Corner cases

+++ src/services/category.service.ts
@@ -1,15 +1,21 @@
 import { CategoryRepository } from '../repositories/category.repository';
 import { Category } from '../entities/category.entity';
 import { CreateCategoryDto, createCategorySchema, UpdateCategoryDto, updateCategorySchema } from '../dto/category';
-import { CATEGORY_ALREADY_EXISTS, CATEGORY_NOT_FOUND } from '../utilities/messages.utility';
+import { CATEGORY_ALREADY_EXISTS, CATEGORY_ALREADY_USED_IN_ROOM, CATEGORY_ALREADY_USED_IN_WORD, CATEGORY_NOT_FOUND } from '../utilities/messages.utility';
 import { mapJoiErrors } from '../middlewares/validation-error.middleware';
+import { RoomRepository } from '../repositories/room.repository';
+import { WordCategoryRepository } from '../repositories/word-category.repository';
 
 
 export class CategoryService {
-    private categoryRepository: CategoryRepository;
-    
+    private categoryRepository      : CategoryRepository;
+    private roomRepository          : RoomRepository;
+    private wordCategoryRepository  : WordCategoryRepository;
+
     constructor() {
-        this.categoryRepository = new CategoryRepository();
+        this.categoryRepository     = new CategoryRepository();
+        this.roomRepository         = new RoomRepository();
+        this.wordCategoryRepository = new WordCategoryRepository();
     }
 
     async getAllCategories(): Promise<Category[]> {
@@ -46,8 +52,12 @@ export class CategoryService {
 
     async deleteCategory(id: number): Promise<void> {
         const responseById = await this.categoryRepository.findCategoryById(id);
+        const responseRoomByIdCategory = await this.roomRepository.findRoomByIdCategory(id);
+        const responseWordByIdCategory = await this.wordCategoryRepository.findWordByIdCategory(id);
 
         if (!responseById) throw new Error(CATEGORY_NOT_FOUND);
+        if( responseRoomByIdCategory.length > 0 ) throw new Error(CATEGORY_ALREADY_USED_IN_ROOM);
+        if( responseWordByIdCategory.length > 0 ) throw new Error(CATEGORY_ALREADY_USED_IN_WORD);
 
         this.categoryRepository.deleteCategory(id);
     }
